+++ packages/grafana-data/src/dataframe/index.ts
@@ -4,10 +4,4 @@ export * from './CircularDataFrame';
 export * from './MutableDataFrame';
 export * from './processDataFrame';
 export * from './dimensions';
-
-// NOTE: We can not export arrow in the global scope because it will crash phantomjs
-// In core, this is loaded async.  In plugins you can import using:
-//
-// import { resultsToDataFrames } from '@grafana/data/dataframe/arrow/ArrowDataFrame'
-//
-// export * from './arrow/ArrowDataFrame';
+export * from './ArrowDataFrame';
